@@ -1,7 +1,7 @@
 project(lifelong)
 
 cmake_minimum_required (VERSION 3.20)
-set(CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_STANDARD 17) # use std::filesystem
 
 IF(NOT CMAKE_BUILD_TYPE)
     #SET(CMAKE_BUILD_TYPE "DEBUG")
@@ -14,7 +14,7 @@ include_directories("inc")
 file(GLOB SOURCES "src/*.cpp")
 
 # Find Boost
-find_package( Boost REQUIRED COMPONENTS program_options system filesystem)
+find_package( Boost REQUIRED COMPONENTS program_options system)
 message(STATUS "Boost_INCLUDE_DIRS: ${Boost_INCLUDE_DIRS}")
 message(STATUS "Boost_LIBRARIES: ${Boost_LIBRARIES}")
 
